requirements.txt
transformer
py7zr
datasets
einops
accelerate
bitsandbytes
evaluate
bert_score
unbabel-comet
absl-py
nltk==3.8.1
rouge-score==0.1.2
sacremoses==0.0.53
sortedcontainers
subword-nmt
google-cloud-storage
scikit-learn-extra
torchmetrics
peft
git+https://github.com/yuchenlin/LLM-Blender.git